@@ -12,72 +12,73 @@
   "private": true,
   "dependencies": {
     "@angular-devkit/build-angular": "^0.803.0",
-    "@angular/animations": "^7.2.3",
-    "@angular/cdk": "^7.3.0",
-    "@angular/cdk-experimental": "^7.3.4",
-    "@angular/common": "~7.2.0",
-    "@angular/compiler": "~7.2.0",
-    "@angular/core": "~7.2.0",
-    "@angular/forms": "~7.2.0",
-    "@angular/material": "^7.3.0",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/router": "~7.2.0",
-    "@types/express": "^4.16.1",
-    "@types/plotly.js": "^1.43.0",
+    "@angular/animations": "^8.2.3",
+    "@angular/cdk": "^8.1.4",
+    "@angular/cdk-experimental": "^8.1.4",
+    "@angular/common": "~8.2.3",
+    "@angular/compiler": "~8.2.3",
+    "@angular/core": "~8.2.3",
+    "@angular/forms": "~8.2.3",
+    "@angular/material": "^8.1.4",
+    "@angular/platform-browser": "~8.2.3",
+    "@angular/platform-browser-dynamic": "~8.2.3",
+    "@angular/router": "~8.2.3",
+    "@types/express": "^4.17.1",
+    "@types/plotly.js": "^1.44.10",
     "@types/socket.io": "^2.1.2",
     "@types/socket.io-client": "^1.4.32",
     "awesome-swiper": "^1.3.0",
-    "convert-csv-to-json": "0.0.13",
-    "core-js": "^2.5.4",
-    "d3": "^5.9.1",
-    "express": "^4.16.4",
+    "convert-csv-to-json": "0.0.15",
+    "core-js": "^3.2.1",
+    "cors": "^2.8.5",
+    "d3": "^5.11.0",
+    "express": "^4.17.1",
     "greiner-hormann": "^1.4.1",
     "http": "0.0.0",
-    "igniteui-angular": "^7.2.3",
-    "js-angusj-clipper": "^1.0.1",
+    "igniteui-angular": "^8.1.5",
+    "js-angusj-clipper": "^1.0.2",
     "js-clipper": "^1.0.1",
     "material-design-icons": "^3.0.1",
     "ng2-dragula": "^2.1.1",
-    "ngx-drag-scroll": "^7.1.4",
-    "ngx-swiper-wrapper": "^7.2.1",
+    "ngx-drag-scroll": "^8.0.0-beta.2",
+    "ngx-swiper-wrapper": "^8.0.2",
     "ngx-useful-swiper": "^8.0.1",
     "process": "^0.11.10",
-    "rxjs": "~6.3.3",
+    "rxjs": "~6.5.2",
     "socket.io": "^2.2.0",
     "socket.io-client": "^2.2.0",
     "swiper": "^4.5.0",
-    "tslib": "^1.9.0",
+    "tslib": "^1.10.0",
     "typeface-titillium-web": "0.0.72",
-    "zone.js": "~0.8.26"
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
     "@angular/cli": "^8.3.0",
-    "@angular/compiler-cli": "~7.2.0",
-    "@angular/language-service": "~7.2.0",
-    "@types/d3": "^5.7.1",
-    "@types/jasmine": "~2.8.8",
+    "@angular/compiler-cli": "~8.2.3",
+    "@angular/language-service": "~8.2.3",
+    "@types/d3": "^5.7.2",
+    "@types/jasmine": "~3.4.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/js-clipper": "0.0.28",
-    "@types/node": "~8.9.4",
+    "@types/node": "~12.7.2",
     "advanced-json-path": "^1.0.8",
     "angular-translate-csv-to-json": "^1.1.0",
-    "codelyzer": "~4.5.0",
-    "gulp": "^4.0.0",
-    "gulp-typescript": "^5.0.0",
-    "jasmine-core": "~2.99.1",
+    "codelyzer": "~5.1.0",
+    "gulp": "^4.0.2",
+    "gulp-typescript": "^5.0.1",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.1.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "ngx-spec": "^2.0.0",
+    "karma": "~4.2.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.0",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "ngx-spec": "^2.1.4",
     "node-sass": "^4.12.0",
     "protractor": "~5.4.0",
-    "style-loader": "^0.23.1",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "^3.2.4"
+    "style-loader": "^1.0.0",
+    "ts-node": "~8.3.0",
+    "tslint": "~5.19.0",
+    "typescript": "^3.5.3"
   }
 }
