@@ -24,7 +24,7 @@ export class WebsocketService {
     console.log("this.socket: ", this.socket);
   }
 
-  private socket = io('http://localhost:3000');
+  private socket = io('http://localhost:3000', { transports : ['websocket'] });
   
   public lockedCM = [{"locked": false, "graphFactor": 0},
                       {"locked": false, "graphFactor": 0},
